fix: generate metadata at the entity layout
