+++ apps/web/app/space/[id]/[entityId]/layout.tsx
@@ -2,6 +2,8 @@ import { SYSTEM_IDS } from '@geogenesis/ids';
 
 import * as React from 'react';
 
+import { Metadata } from 'next';
+
 import { Environment } from '~/core/environment';
 import { API, Subgraph } from '~/core/io';
 import { fetchEntityType } from '~/core/io/fetch-entity-type';
@@ -9,7 +11,7 @@ import { EntityStoreProvider } from '~/core/state/entity-page-store/entity-store
 import { DEFAULT_PAGE_SIZE } from '~/core/state/triple-store/triple-store';
 import { Entity as IEntity, Triple } from '~/core/types';
 import { Entity } from '~/core/utils/entity';
-import { NavUtils } from '~/core/utils/utils';
+import { NavUtils, getOpenGraphMetadataForEntity } from '~/core/utils/utils';
 import { Value } from '~/core/utils/value';
 
 import { Spacer } from '~/design-system/spacer';
@@ -26,6 +28,41 @@ const TABS = ['Overview', 'Activity'] as const;
 interface Props {
   params: { id: string; entityId: string };
   children: React.ReactNode;
+}
+
+export const runtime = 'edge';
+
+export async function generateMetadata({ params }: Props): Promise<Metadata> {
+  const spaceId = params.id;
+  const entityId = params.entityId;
+  const config = Environment.getConfig(process.env.NEXT_PUBLIC_APP_ENV);
+
+  const entity = await Subgraph.fetchEntity({ endpoint: config.subgraph, id: entityId });
+  const { entityName, description, openGraphImageUrl } = getOpenGraphMetadataForEntity(entity);
+
+  return {
+    title: entityName ?? 'New entity',
+    description,
+    openGraph: {
+      title: entityName ?? 'New entity',
+      description,
+      url: `https://geobrowser.io${NavUtils.toEntity(spaceId, entityId)}`,
+      images: [
+        {
+          url: openGraphImageUrl,
+        },
+      ],
+    },
+    twitter: {
+      card: 'summary_large_image',
+      description,
+      images: [
+        {
+          url: openGraphImageUrl,
+        },
+      ],
+    },
+  };
 }
 
 export default async function ProfileLayout({ children, params }: Props) {
